+++ docs/RELEASE-NOTES.md
@@ -8,6 +8,7 @@ Breaking changes:
 New features:
 - Added `UriSlugBuilder` class.
 - Added `UriExtensions.GetPathSegments` method.
+- Added `UriExtensions.GetRoot` method.
 
 Bug fixes / internal changes:
 - Updated `ByteDev.Strings` dependency to version `11.0.0`.